requirements.txt.txt
Flask
google-api-python-client
google-auth
google-auth-oauthlib
geopy
beautifulsoup4
requests
